Use uintmax_t for ino_t.

+++ usr.sbin/pstat/pstat.c
@@ -33,7 +33,7 @@
  * @(#) Copyright (c) 1980, 1991, 1993, 1994 The Regents of the University of California.  All rights reserved.
  * @(#)pstat.c	8.16 (Berkeley) 5/9/95
  * $FreeBSD: src/usr.sbin/pstat/pstat.c,v 1.49.2.5 2002/07/12 09:12:49 des Exp $
- * $DragonFly: src/usr.sbin/pstat/pstat.c,v 1.21 2007/05/09 04:33:52 dillon Exp $
+ * $DragonFly: src/usr.sbin/pstat/pstat.c,v 1.22 2008/03/04 18:10:44 swildner Exp $
  */
 
 #define _KERNEL_STRUCTURES
@@ -558,7 +558,7 @@ nfs_print(struct vnode *vp)
 	*flags = '\0';
 
 #define VT	np->n_vattr
-	printf(" %6ld %5s", VT.va_fileid, flagbuf);
+	printf(" %6ju %5s", (uintmax_t)VT.va_fileid, flagbuf);
 	type = VT.va_mode & S_IFMT;
 	if (S_ISCHR(VT.va_mode) || S_ISBLK(VT.va_mode))
 		if (usenumflag || ((name = devname((VT.va_rmajor << 8) | VT.va_rminor, type)) == NULL))
